chore: move mysql transformer to `type_mapper.go`
