+++ app/server/datasource/rdbms/mysql/type_mapper.go
@@ -1,10 +1,17 @@
 package mysql
 
 import (
+	"errors"
+	"fmt"
+
+	"github.com/apache/arrow/go/v13/arrow/array"
+	"github.com/go-mysql-org/go-mysql/mysql"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
 
 	api_service_protos "github.com/ydb-platform/fq-connector-go/api/service/protos"
+	"github.com/ydb-platform/fq-connector-go/app/server/conversion"
 	"github.com/ydb-platform/fq-connector-go/app/server/datasource"
+	"github.com/ydb-platform/fq-connector-go/app/server/paging"
 	"github.com/ydb-platform/fq-connector-go/common"
 )
 
@@ -21,17 +28,63 @@ func (typeMapper) SQLTypeToYDBColumn(columnName, _ string, _ *api_service_protos
 
 func NewTypeMapper() datasource.TypeMapper { return typeMapper{} }
 
-// func transformerFromTypeNames(typeNames []string, ydbTypes []*Ydb.Type, cc conversion.Collection) (paging.RowTransformer[any], error) {
-// 	acceptors := make([]any, 0, len(typeNames))
-// 	appenders := make([]func(acceptor any, builder array.Builder))
-//
-// 	for i, name := range typeNames {
-// 		switch name {
-// 		default:
-// 			acceptors = append(acceptors, "int")
-// 			appenders = append(appenders, func(acceptor any, builder array.Builder) error {
-// 				cast := acceptor.(*mysql.)
-// 			})
-// 		}
-// 	}
-// }
+func transformerFromTypeIDs(ids []uint8, _ []*Ydb.Type, cc conversion.Collection) (paging.RowTransformer[any], error) {
+	acceptors := make([]any, 0, len(ids))
+	appenders := make([]func(acceptor any, builder array.Builder) error, 0, len(ids))
+
+	for _, id := range ids {
+		switch id {
+		case mysql.MYSQL_TYPE_LONG:
+			acceptors = append(acceptors, new(*int32))
+			appenders = append(appenders, makeAppender[int32, int32, *array.Int32Builder](cc.Int32()))
+		default:
+			panic(fmt.Sprintf("Type %d not implemented yet!", id))
+		}
+	}
+
+	return paging.NewRowTransformer(acceptors, appenders, nil), nil
+}
+
+func makeAppender[
+	IN common.ValueType,
+	OUT common.ValueType,
+	AB common.ArrowBuilder[OUT],
+](conv conversion.ValueConverter[IN, OUT]) func(acceptor any, builder array.Builder) error {
+	return func(acceptor any, builder array.Builder) error {
+		return appendValueToArrowBuilder[IN, OUT, AB](acceptor, builder, conv)
+	}
+}
+
+func appendValueToArrowBuilder[IN common.ValueType, OUT common.ValueType, AB common.ArrowBuilder[OUT]](
+	acceptor any,
+	builder array.Builder,
+	conv conversion.ValueConverter[IN, OUT],
+) error {
+	cast := acceptor.(**IN)
+
+	if *cast == nil {
+		builder.AppendNull()
+
+		return nil
+	}
+
+	value := **cast
+
+	out, err := conv.Convert(value)
+	if err != nil {
+		if errors.Is(err, common.ErrValueOutOfTypeBounds) {
+			builder.AppendNull()
+
+			return nil
+		}
+
+		return fmt.Errorf("convert value %v: %w", value, err)
+	}
+
+	//nolint:forcetypeassert
+	builder.(AB).Append(out)
+
+	*cast = nil
+
+	return nil
+}
